skip some crashing tests when running against GDAL < 1.9.0
git-svn-id: https://svn.osgeo.org/gdal/trunk@23540 f0d54148-0727-0410-94bb-9a71ac55c965

+++ autotest/ogr/ogr_gml_read.py
@@ -1014,6 +1014,10 @@ def ogr_gml_25():
     if not gdaltest.have_gml_reader:
         return 'skip'
 
+    if int(gdal.VersionInfo('VERSION_NUM')) < 1900:
+        gdaltest.post_reason('would crash')
+        return 'skip'
+
     try:
         os.remove( 'data/curveProperty.gfs' )
     except:
